fix: resolve all 9 mypy errors across musaeus/
- fuzzy.py: wrap rapidfuzz/difflib returns in float()
- db.py: wrap fetchone()[0] in int()
- loudness.py: explicit dict[str, Any] annotations
- hasher.py: guard proc.stderr for None before .read()
- rebuild.py: extract errors list with explicit type
- stages/scholar.py: annotate json.loads result

Result: mypy musaeus/ → 0 errors (33 files checked)

+++ musaeus/fuzzy.py
@@ -87,10 +87,10 @@ def similarity(a: str, b: str, pre_normalized: bool = False) -> float:
         b = normalize(b)
 
     if _HAVE_RAPIDFUZZ:
-        return _fuzz.ratio(a, b)
+        return float(_fuzz.ratio(a, b))
     else:
         # difflib returns 0..1; scale to 0..100
-        return _difflib.SequenceMatcher(None, a, b).ratio() * 100
+        return float(_difflib.SequenceMatcher(None, a, b).ratio() * 100)
 
 
 def token_similarity(a: str, b: str, pre_normalized: bool = False) -> float:
@@ -105,7 +105,7 @@ def token_similarity(a: str, b: str, pre_normalized: bool = False) -> float:
         b = normalize(b)
 
     if _HAVE_RAPIDFUZZ:
-        return _fuzz.token_sort_ratio(a, b)
+        return float(_fuzz.token_sort_ratio(a, b))
     return similarity(a, b, pre_normalized=True)
 
 
@@ -119,7 +119,7 @@ def partial_similarity(a: str, b: str, pre_normalized: bool = False) -> float:
         b = normalize(b)
 
     if _HAVE_RAPIDFUZZ:
-        return _fuzz.partial_ratio(a, b)
+        return float(_fuzz.partial_ratio(a, b))
     return similarity(a, b, pre_normalized=True)
